+++ Contents/mods/Evolving Traits World/42.19/media/lua/shared/ETW_ModData.lua
@@ -104,6 +104,7 @@ function ETW_ModData.createETWModData(playerIndex, player)
 	modData.PainToleranceCounter = modData.PainToleranceCounter or 0
 	modData.UniqueClothingRipped = modData.UniqueClothingRipped or {}
 	modData.ImmunitySystemCounter = modData.ImmunitySystemCounter or 0
+	modData.PagesReadCounter = modData.PagesReadCounter or 0
 
 	modData.MentalStateInLast60Min = modData.MentalStateInLast60Min or { 0.75 }
 	modData.MentalStateInLast24Hours = modData.MentalStateInLast24Hours or { 0.75 }